+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Reword log messages that implied duplicate uploads
   [#296](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/296)
 
+* Fix bugsnag.sharedObjectPaths not uploading SO files
+  [#295](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/295)
+
 * Only set mappingFilesProvider on release task if obfuscation enabled
   [#292](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/292)