+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## TBD
+## 5.1.0 (2020-09-17)
 
 * Remove uploadDebugBuildMappings flag in favour of variantFilter
   [#301](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/301)